third_party/WebKit/LayoutTests/transforms/webkit-perspective-origin-csstext.html

<!doctype html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function(){
    var div = document.createElement('div');
    div.style.webkitPerspectiveOrigin = '20px 30px';
    assert_equals(div.style.cssText, 'perspective-origin: 20px 30px;');
}, "cssText behaves as expected with -webkit-perspective-origin");
</script>
